requirements.txt
aiohttp>=3.9,<4.0
pyyaml>=6.0,<7.0
httpx>=0.27,<1.0

# AI Providers (install the one you use)
# pip install google-generativeai  # For Gemini (default)
# pip install anthropic             # For Anthropic Claude
# pip install openai                # For OpenAI

# Optional: for automated scanning
# pip install tweepy                # Twitter API
# pip install feedparser            # RSS feeds

# Testing
pytest>=8.0
pytest-asyncio>=0.23
pytest-aiohttp>=1.0
